api/requirements.txt
asgiref==3.3.4
blis==0.7.4
catalogue==2.0.3
certifi==2020.12.5
chardet
click
contextvars
cymem
Cython
dataclasses
Django
django-cors-headers
djangorestframework
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
gunicorn==20.1.0
idna==2.10
immutables==0.15
importlib-metadata==3.10.1
Jinja2==2.11.3
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.19.5
packaging==20.9
# pathy==0.4.0
# pkg-resources==0.0.0
preshed==3.0.5
pydantic==1.7.3
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.2
sqlparse==0.4.1
srsly==2.4.1
thinc==8.0.2
tqdm==4.60.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
whitenoise==5.2.0
zipp==3.4.1
